# docs/swagger.yaml
basePath: /api
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
...
  models.ErrorResponse:
    properties:
      created_at:
        type: string
      error:
        type: string
    type: object
  models.ForgotPasswordInput:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  models.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        properties:
          email:
            type: string
          id:
            type: string
          name:
            type: string
          role:
            type: string
        type: object
    type: object
  models.ResetPasswordInput:
    properties:
      new_password:
        example: MyNewPassword123
        type: string
      token:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: RESTful API for signing and verifying files with digital signatures
  title: Tawtheeq API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a reset password link to the user's email if it exists
      parameters:
      - description: User email
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Send password reset email
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user and return JWT token
      parameters:
      - description: Login Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Login
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password if the token is valid
      parameters:
      - description: Token and new password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Reset password using token
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"